+++ src/FESpaces/FEAutodiff.jl
@@ -16,11 +16,13 @@ end
 
 function _gradient(f,uh,fuh::DomainContribution;tag::GridapADTag=get_ad_level(fuh)+1)
   terms = DomainContribution(;ad_level = tag)
+  V = get_fe_space(uh)
+  T = eltype(get_vector_type(V))
   for trian in get_domains(fuh)
     g = _change_argument(gradient,f,trian,uh)
     cell_u = get_cell_dof_values(uh)
     cell_id = _compute_cell_ids(uh,trian)
-    cell_grad = autodiff_array_gradient(g,cell_u,cell_id;tag)
+    cell_grad = autodiff_array_gradient(T,g,cell_u,cell_id;tag)
     add_contribution!(terms,trian,cell_grad)
   end
   terms
@@ -42,11 +44,17 @@ end
 
 function _jacobian(f,uh,fuh::DomainContribution;tag::GridapADTag=get_ad_level(fuh)+1)
   terms = DomainContribution(;ad_level=tag)
+  V = get_fe_space(uh)
+  T = eltype(get_vector_type(V))
+  # For now, taking the Jacobian of a complex-valued functional is not implemented.
+  # To implement this, we need to consider the case of f being either holomorphic or
+  # non-holomorphic
+  T<:Complex && @notimplemented "Jacobian's of complex-valued functionals are not yet implemented"
   for trian in get_domains(fuh)
     g = _change_argument(jacobian,f,trian,uh)
     cell_u = get_cell_dof_values(uh)
     cell_id = _compute_cell_ids(uh,trian)
-    cell_grad = autodiff_array_jacobian(g,cell_u,cell_id;tag)
+    cell_grad = autodiff_array_jacobian(T,g,cell_u,cell_id;tag)
     add_contribution!(terms,trian,cell_grad)
   end
   terms
@@ -71,11 +79,14 @@ end
 
 function _hessian(f,uh,fuh::DomainContribution;tag::GridapADTag=get_ad_level(fuh)+1)
   terms = DomainContribution(;ad_level = tag+1) # Two levels consumed
+  V = get_fe_space(uh)
+  T = eltype(get_vector_type(V))
+  T<:Complex && @notimplemented "Hessian's of complex-valued functionals are not yet implemented"
   for trian in get_domains(fuh)
     g = _change_argument(hessian,f,trian,uh)
     cell_u = get_cell_dof_values(uh)
     cell_id = _compute_cell_ids(uh,trian)
-    cell_grad = autodiff_array_hessian(g,cell_u,cell_id;tag)
+    cell_grad = autodiff_array_hessian(T,g,cell_u,cell_id;tag)
     add_contribution!(terms,trian,cell_grad)
   end
   terms
@@ -148,7 +159,7 @@ end
 # which returns ydual_θ = df/duᶿ for θ ∈ {+, -}
 # We them merge them into a 2-block BlockVector, so that we obtain
 #   result = [df/du⁺, df/du⁻]
-function Arrays.autodiff_array_gradient(a, i_to_x, j_to_i::SkeletonPair; tag=default_tag(ForwardDiff.gradient,a))
+function Arrays.autodiff_array_gradient(V, a, i_to_x, j_to_i::SkeletonPair; tag=default_tag(ForwardDiff.gradient,a))
   i_to_cfg = lazy_map(ConfigMap(ForwardDiff.gradient,tag),i_to_x)
   i_to_xdual = lazy_map(DualizeMap(),i_to_cfg,i_to_x)
 
@@ -163,6 +174,10 @@ function Arrays.autodiff_array_gradient(a, i_to_x, j_to_i::SkeletonPair; tag=def
   j_to_cfg_minus = Arrays.autodiff_array_reindex(i_to_cfg,j_to_i.minus)
   j_to_result_minus = lazy_map(AutoDiffMap(),j_to_cfg_minus,j_to_ydual_minus)
 
+  return _skeleton_autodiff_merge_gradient(j_to_result_plus,j_to_result_minus)
+end
+
+function _skeleton_autodiff_merge_gradient(j_to_result_plus,j_to_result_minus)
   # Assemble on SkeletonTriangulation expects an array of interior of facets
   # where each entry is a 2-block BlockVector with the first block being the
   # contribution of the plus side and the second, the one of the minus side
@@ -176,7 +191,7 @@ end
 # We them merge them as columns into a 2x2 block matrix, so that we obtain
 # ydual = [dr⁺/du⁺ dr⁺/du⁻] = [ydual_plus, ydual_minus]
 #         [dr⁻/du⁺ dr⁻/du⁻]
-function Arrays.autodiff_array_jacobian(a, i_to_x, j_to_i::SkeletonPair; tag=default_tag(ForwardDiff.jacobian,a))
+function Arrays.autodiff_array_jacobian(V, a, i_to_x, j_to_i::SkeletonPair; tag=default_tag(ForwardDiff.jacobian,a))
   i_to_cfg = lazy_map(ConfigMap(ForwardDiff.jacobian,tag),i_to_x)
   i_to_xdual = lazy_map(DualizeMap(),i_to_cfg,i_to_x)
 
@@ -191,8 +206,11 @@ function Arrays.autodiff_array_jacobian(a, i_to_x, j_to_i::SkeletonPair; tag=def
   j_to_cfg_minus = Arrays.autodiff_array_reindex(i_to_cfg,j_to_i.minus)
   j_to_result_minus = lazy_map(AutoDiffMap(),j_to_cfg_minus,j_to_ydual_minus)
 
+  return _skeleton_autodiff_merge_jacobian(j_to_result_plus,j_to_result_minus)
+end
+
+function _skeleton_autodiff_merge_jacobian(j_to_result_plus,j_to_result_minus)
   # Merge the columns into a 2x2 block matrix
-  # I = [[(CartesianIndex(i,),CartesianIndex(i,j)) for i in 1:2] for j in 1:2]
   I = [
     [(CartesianIndex(1,), CartesianIndex(1, 1)), (CartesianIndex(2,), CartesianIndex(2, 1))], # Plus  -> First column
     [(CartesianIndex(1,), CartesianIndex(1, 2)), (CartesianIndex(2,), CartesianIndex(2, 2))]  # Minus -> Second column
@@ -200,4 +218,29 @@ function Arrays.autodiff_array_jacobian(a, i_to_x, j_to_i::SkeletonPair; tag=def
   is_single_field = eltype(eltype(j_to_result_plus)) <: AbstractArray
   k = is_single_field ? Arrays.MergeBlockMap((2,2),I) : Arrays.BlockBroadcasting(Arrays.MergeBlockMap((2,2),I))
   lazy_map(k,j_to_result_plus,j_to_result_minus)
+end
+
+# Skeleton + Complex (gradient)
+# Split a into real and imaginary components and take gradients sperately.
+function Arrays.autodiff_array_gradient(::Type{<:Complex},a,i_to_x,j_to_i::SkeletonPair;tag=default_tag(ForwardDiff.gradient,a))
+  s = lazy_map(Broadcasting(imag),i_to_x)
+  r = lazy_map(Broadcasting(real),i_to_x)
+  i_to_cfg_r = lazy_map(ConfigMap(ForwardDiff.gradient,tag),r)
+  i_to_rdual = lazy_map(DualizeMap(),i_to_cfg_r,r)
+  j_to_yrdual_plus,j_to_yrdual_minus = a(lazy_map((r,s) -> r + im*s,i_to_rdual,s))
+  i_to_cfg_s = lazy_map(ConfigMap(ForwardDiff.gradient,tag),s)
+  i_to_sdual = lazy_map(DualizeMap(),i_to_cfg_s,s)
+  j_to_ysdual_plus,j_to_ysdual_minus = a(lazy_map((r,s) -> r + im*s,r,i_to_sdual))
+  j_to_cfg_r_plus = Arrays.autodiff_array_reindex(i_to_cfg_r,j_to_i.plus)
+  j_to_cfg_r_minus = Arrays.autodiff_array_reindex(i_to_cfg_r,j_to_i.minus)
+  j_to_cfg_s_plus = Arrays.autodiff_array_reindex(i_to_cfg_s,j_to_i.plus)
+  j_to_cfg_s_minus = Arrays.autodiff_array_reindex(i_to_cfg_s,j_to_i.minus)
+  j_to_result_plus = lazy_map((r,s) -> r + im*s,
+    lazy_map(AutoDiffMap(),j_to_cfg_r_plus,lazy_map(Broadcasting(real),j_to_yrdual_plus)),
+    lazy_map(AutoDiffMap(),j_to_cfg_s_plus,lazy_map(Broadcasting(real),j_to_ysdual_plus)))
+  j_to_result_minus = lazy_map((r,s) -> r + im*s,
+    lazy_map(AutoDiffMap(),j_to_cfg_r_minus,lazy_map(Broadcasting(real),j_to_yrdual_minus)),
+    lazy_map(AutoDiffMap(),j_to_cfg_s_minus,lazy_map(Broadcasting(real),j_to_ysdual_minus)))
+    
+  return _skeleton_autodiff_merge_gradient(j_to_result_plus,j_to_result_minus)
 end